Start refactoring user table
